Refactored to remove redundant ContentMapperSubscriber
