+++ src/Sulu/Bundle/ContentBundle/Resources/config/search.xml
@@ -6,7 +6,7 @@
     <parameters>
         <parameter key="sulu_content.search.metadata.provider.structure.class">Sulu\Bundle\ContentBundle\Search\Metadata\StructureProvider</parameter>
         <parameter key="sulu_content.search.event_subscriber.blame_timestamp.class">Sulu\Bundle\ContentBundle\Search\EventSubscriber\BlameTimestampSubscriber</parameter>
-        <parameter key="sulu_content.search.event_listener.search.class">Sulu\Bundle\ContentBundle\Search\EventListener\SearchListener</parameter>
+        <parameter key="sulu_content.search.event_subscriber.structure.class">Sulu\Bundle\ContentBundle\Search\EventSubscriber\StructureSubscriber</parameter>
         <parameter key="sulu_content.search.event_listener.reindex.class">Sulu\Bundle\ContentBundle\Search\EventListener\ReindexListener</parameter>
     </parameters>
 
@@ -27,12 +27,6 @@
             <tag name="kernel.event_subscriber" />
         </service>
 
-        <!-- Search event to be sure that all structures are loaded (generated) -->
-        <service id="sulu_content.search.event_listener.search" class="%sulu_search.event_listener.search.class%">
-            <argument type="service" id="sulu.content.structure_manager" />
-            <tag name="kernel.event_listener" event="massive_search.search" method="onSearch" />
-        </service>
-
         <!-- Rebuild structure index on massive:search:index:rebuild -->
         <service id="sulu_content.search.event_listener.search" class="%sulu_content.search.event_listener.reindex.class%">
             <argument type="service" id="sulu_document_manager.document_manager" />
@@ -40,6 +34,12 @@
             <argument type="service" id="massive_search.search_manager" />
             <argument>%sulu_content.search.mapping%</argument>
             <tag name="kernel.event_listener" event="massive_search.index_rebuild" method="onIndexRebuild" />
+        </service>
+
+        <service id="sulu_content.search.event_subscriber.structure" class="%sulu_content.search.event_subscriber.structure.class%">
+            <argument type="service" id="massive_search.search_manager" />
+            <argument type="service" id="sulu_document_manager.metadata_factory" />
+            <tag name="sulu_document_manager.event_subscriber" />
         </service>
 
     </services>
