+++ src/Sulu/Bundle/SearchBundle/Resources/config/search.xml
@@ -7,7 +7,6 @@
         <parameter key="sulu_search.controller.search.class">Sulu\Bundle\SearchBundle\Controller\SearchController</parameter>
         <parameter key="sulu_search.search.factory.class">Sulu\Bundle\SearchBundle\Search\Factory</parameter>
         <parameter key="sulu_search.admin.class">Sulu\Bundle\SearchBundle\Admin\SuluSearchAdmin</parameter>
-        <parameter key="sulu_search.event_subscriber.content.class">Sulu\Bundle\SearchBundle\EventListener\StructureSubscriber</parameter>
     </parameters>
 
     <services>
@@ -28,10 +27,5 @@
             <argument>%sulu_admin.name%</argument>
         </service>
 
-        <service id="sulu_search.event_subscriber.content" class="%sulu_search.event_subscriber.content.class%">
-            <argument type="service" id="massive_search.search_manager" />
-            <argument type="service" id="sulu_document_manager.metadata_factory" />
-            <tag name="sulu_document_manager.event_subscriber" />
-        </service>
     </services>
 </container>
